--- a/src/BlockArrays.jl
+++ b/src/BlockArrays.jl
@@ -16,6 +16,8 @@ export undef_blocks, undef, findblock, findblockindex
 
 export khatri_rao
 
+export blockappend!
+
 import Base: @propagate_inbounds, Array, to_indices, to_index,
             unsafe_indices, first, last, size, length, unsafe_length,
             unsafe_convert,