--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -9,4 +9,6 @@ using BlockArrays, LinearAlgebra, Test
     include("test_blockbroadcast.jl")
     include("test_blocklinalg.jl")
     include("test_blockproduct.jl")
+    include("test_blockreduce.jl")
+    include("test_blockdeque.jl")
 end