… docs/src/lib/public.md
… test/test_blockreduce.jl