This PR adds `append!`, `push!`, `pushfirst!`, `pop!`, and `popfirst!` for `BlockVector`.

It also adds `foldl` (and `reduce` because why not) for `BlockVector` since it is essential for fast iteration (see https://github.com/JuliaArrays/BlockArrays.jl/pull/116#discussion_r426124156 below).

---

It's not the main point of this PR but a few quick benchmarks for `foldl` and `reduce` alone (just to help justify their inclusion):

Before:

```julia
julia> @btime sum($(mortar([rand(100) for _ in 1:10])));
  13.494 μs (0 allocations: 0 bytes)

julia> @btime foldl(+, $(mortar([rand(100) for _ in 1:10])));
  13.637 μs (0 allocations: 0 bytes)
```

After:

```julia
julia> @btime sum($(mortar([rand(100) for _ in 1:10])));
  130.989 ns (0 allocations: 0 bytes)

julia> @btime foldl(+, $(mortar([rand(100) for _ in 1:10])));
  1.090 μs (3 allocations: 80 bytes)
```